Revert "Refactor: separate SplittableTruncateSizedRestrictions" (#35700)
