+++ sdks/java/harness/src/main/java/org/apache/beam/fn/harness/SplitResultsWithStopIndex.java
@@ -20,7 +20,6 @@ package org.apache.beam.fn.harness;
 import com.google.auto.value.AutoValue;
 import org.apache.beam.sdk.annotations.Internal;
 import org.checkerframework.checker.nullness.qual.Nullable;
-import org.checkerframework.dataflow.qual.Pure;
 
 @AutoValue
 @AutoValue.CopyAnnotations
@@ -28,18 +27,15 @@ import org.checkerframework.dataflow.qual.Pure;
 abstract class SplitResultsWithStopIndex {
   public static SplitResultsWithStopIndex of(
       WindowedSplitResult windowSplit,
-      HandlesSplits.@Nullable SplitResult downstreamSplit,
+      HandlesSplits.SplitResult downstreamSplit,
       int newWindowStopIndex) {
     return new AutoValue_SplitResultsWithStopIndex(
         windowSplit, downstreamSplit, newWindowStopIndex);
   }
 
-  @Pure
-  public abstract WindowedSplitResult getWindowSplit();
+  public abstract @Nullable WindowedSplitResult getWindowSplit();
 
-  @Pure
   public abstract HandlesSplits.@Nullable SplitResult getDownstreamSplit();
 
-  @Pure
   public abstract int getNewWindowStopIndex();
 }
