+++ sdks/java/harness/src/main/java/org/apache/beam/fn/harness/WindowedSplitResult.java
@@ -21,7 +21,6 @@ import com.google.auto.value.AutoValue;
 import org.apache.beam.sdk.annotations.Internal;
 import org.apache.beam.sdk.values.WindowedValue;
 import org.checkerframework.checker.nullness.qual.Nullable;
-import org.checkerframework.dataflow.qual.Pure;
 
 /** Internal class to hold the primary and residual roots when converted to an input element. */
 @AutoValue
@@ -29,10 +28,10 @@ import org.checkerframework.dataflow.qual.Pure;
 @Internal
 abstract class WindowedSplitResult {
   public static WindowedSplitResult forRoots(
-      @Nullable WindowedValue<?> primaryInFullyProcessedWindowsRoot,
-      @Nullable WindowedValue<?> primarySplitRoot,
-      @Nullable WindowedValue<?> residualSplitRoot,
-      @Nullable WindowedValue<?> residualInUnprocessedWindowsRoot) {
+      WindowedValue<?> primaryInFullyProcessedWindowsRoot,
+      WindowedValue<?> primarySplitRoot,
+      WindowedValue<?> residualSplitRoot,
+      WindowedValue<?> residualInUnprocessedWindowsRoot) {
     return new AutoValue_WindowedSplitResult(
         primaryInFullyProcessedWindowsRoot,
         primarySplitRoot,
@@ -40,15 +39,11 @@ abstract class WindowedSplitResult {
         residualInUnprocessedWindowsRoot);
   }
 
-  @Pure
   public abstract @Nullable WindowedValue<?> getPrimaryInFullyProcessedWindowsRoot();
 
-  @Pure
   public abstract @Nullable WindowedValue<?> getPrimarySplitRoot();
 
-  @Pure
   public abstract @Nullable WindowedValue<?> getResidualSplitRoot();
 
-  @Pure
   public abstract @Nullable WindowedValue<?> getResidualInUnprocessedWindowsRoot();
 }
